+++ R/globals.R
@@ -1,1 +1,2 @@
-globalVariables(c(".", "gutenberg_id", "language", "has_text", "gutenberg_metadata"))
+globalVariables(c(".", "gutenberg_id", "language", "has_text",
+                  "gutenberg_metadata", "number", "total"))