+++ R/gutenberg_works.R
@@ -9,8 +9,7 @@
 #'
 #' @param ... Additional filters, given as expressions using the variables
 #' in the \link{gutenberg_metadata} dataset (e.g. \code{author == "Austen, Jane"})
-#' @param languages Vector of languages to include (note that it will not
-#' return cases with multiple languages unless they are specified)
+#' @param languages Vector of languages to include
 #' @param only_text Whether the works must have Gutenberg text attached. Works
 #' without text (e.g. audiobooks) cannot be downloaded with
 #' \code{\link{gutenberg_download}}
@@ -19,6 +18,13 @@
 #' @param distinct Whether to return only one distinct combination of each
 #' title and gutenberg_author_id. If multiple occur (that fulfill the other
 #' conditions), it uses the one with the lowest ID
+#' @param all_languages Whether, if multiple languages are given, all of them
+#' need to be present in a work. For example, if \code{c("en", "fr")} are given,
+#' whether only \code{en/fr} as opposed to English or French works should be
+#' returned
+#' @param only_languages Whether to exclude works that have other languages
+#' besides the ones provided. For example, whether to include \code{en/fr}
+#' when English works are requested
 #'
 #' @return A tbl_df (see the tibble or dplyr packages) with one row for
 #' each work, in the same format as \link{gutenberg_metadata}.
@@ -34,6 +40,8 @@
 #'
 #' @examples
 #'
+#' library(dplyr)
+#'
 #' gutenberg_works()
 #'
 #' # filter conditions
@@ -41,13 +49,28 @@
 #'
 #' # changing default options
 #' gutenberg_works(rights = NULL)
-#' gutenberg_works(languages = "de")
+#'
+#' # language specifications
+#'
+#' gutenberg_works(languages = "es") %>%
+#'   count(language, sort = TRUE)
+#'
+#' gutenberg_works(languages = c("en", "es")) %>%
+#'   count(language, sort = TRUE)
+#'
+#' gutenberg_works(languages = c("en", "es"), all_languages = TRUE) %>%
+#'   count(language, sort = TRUE)
+#'
+#' gutenberg_works(languages = c("en", "es"), only_languages = FALSE) %>%
+#'   count(language, sort = TRUE)
 #'
 #' @export
 gutenberg_works <- function(..., languages = "en",
                             only_text = TRUE,
                             rights = c("Public domain in the USA.", "None"),
-                            distinct = TRUE) {
+                            distinct = TRUE,
+                            all_languages = FALSE,
+                            only_languages = TRUE) {
   utils::data("gutenberg_metadata", package = "gutenbergr", envir = environment())
 
   dots <- lazyeval::lazy_dots(...)
@@ -59,7 +82,30 @@ gutenberg_works <- function(..., languages = "en",
   ret <- filter_(gutenberg_metadata, .dots = dots)
 
   if (!is.null(languages)) {
-    ret <- filter(ret, language %in% languages)
+    lang_spl <- ret %>%
+      select(gutenberg_id, language) %>%
+      tidyr::unnest(language = stringr::str_split(language, "/")) %>%
+      group_by(gutenberg_id) %>%
+      mutate(total = n()) %>%
+      ungroup()
+
+    lang_filt <- lang_spl %>%
+      filter(language %in% languages) %>%
+      group_by(gutenberg_id) %>%
+      mutate(number = n()) %>%
+      ungroup()
+
+    if (all_languages) {
+      lang_filt <- lang_filt %>%
+        filter(number >= length(languages))
+    }
+    if (only_languages) {
+      lang_filt <- lang_filt %>%
+        filter(total == number)
+    }
+
+    ret <- ret %>%
+      filter(gutenberg_id %in% lang_filt$gutenberg_id)
   }
 
   if (!is.null(rights)) {