+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * The license was changed from MIT to GPL-2. This is based on the realization that the catalog data is licensed under the GPL, and the package includes a processed version of the catalog data. (See [here](https://www.gutenberg.org/wiki/Gutenberg:Feeds#The_Complete_Project_Gutenberg_Catalog)).
 * Updated datasets to 5/5/2016 and added a "date_updated" attriute to tell when they were last updated
+* Added `all_languages` and `only_languages` arguments to `gutenberg_works`, allowing fine-grained control of languages. (For example, "either English or French" or "both English and French")
 * Various documentation, vignette, and README adjustments in response to ROpenSci feedback.
 * Added AppVeyor for Windows continuous integration
 * Added code coverage information through codecov.io and covr